backend/requirements.txt
django==3.2.0
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
drf-spectacular==0.21.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
pytest==7.1.1
pyrsistent==0.18.1
psycopg2-binary==2.9.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
webcolors==1.11.1
zipp==3.7.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
isort==5.9.3
tqdm
python-dotenv==0.20.0